EVEREST-596 Added useDbActions hook
